refactor(lens): productionize context lens client
Split the 1500-line ContextLensPanel POC into focused modules under
Channel/ContextLens/ and harden it for production use:

- Gate behind a contextLens feature flag plus gateway URL/token settings
  (new keyValue storage items); remove the raw localStorage read and
  NODE_ENV localhost fallback
- Replace EventSource with a fetch-streaming SSE client so the bearer
  token travels in an Authorization header, with Last-Event-ID replay
  on reconnect and a ref-counted shared connection across Channel and
  PostScreenView
- Replace the dark-only rgba/hex palette with semantic Tamagui theme
  tokens so light mode renders correctly
- Fix impure Date.now() in the run timeline (tick only while a run is
  active), abort in-flight /run lookups on unmount, and drop a no-op
  lookup-status branch
- Add botShip to the tlon-context-lens blob schema with malformed-
  payload tests and docs/post-blobs.md coverage
