+++ packages/app/ui/components/Channel/ContextLens/index.ts
@@ -1,0 +1,15 @@
+export { ContextLensPanel } from './ContextLensPanel';
+export {
+  type ContextLens,
+  type ContextLensEvent,
+  type ContextLensSelectedMessage,
+  type LensStreamState,
+  isContextLensEventActive,
+} from './types';
+export {
+  useContextLensAvailable,
+  useContextLensController,
+  useContextLensEvents,
+  useContextLensGatewayConfig,
+  useContextLensRuns,
+} from './useContextLensStore';
